Move ZoopListItem content to Zoop component
